refactor: Extract HookManager concerns into separate classes
Extract permission generation and file operations from HookManager into
dedicated classes to improve maintainability and separation of concerns.

- Create PermissionGenerator class for Claude Code permission handling
- Create HookFileManager class for filesystem operations
- Reduce HookManager from 763 lines to ~600 lines
- Maintain same public API and functionality
- Keep valuable validation logic in HookManager
