+++ .memento/tickets/done/cleanup-hookmanager-refactor.md
